test: move modular-form dispatch coverage to integration
