+++ tests/math/modular_forms/test_level_one_q_expansions.py
@@ -6,8 +6,6 @@ import pytest
 from pydantic import ValidationError
 
 from jacobian._exact import CanonicalRational
-from jacobian.catalog.catalog import Catalog
-from jacobian.dispatch import invoke_operation
 from jacobian.math.formal_power_series._operations import (
     compute_power,
     compute_scalar_multiply,
@@ -90,41 +88,3 @@ def test_value_rejects_forged_coefficient_during_replay() -> None:
     )
     with pytest.raises(ValidationError, match="does not match"):
         LevelOneModularQExpansion.model_validate(payload)
-
-
-def test_dispatch_returns_the_canonical_typed_q_expansion() -> None:
-    result = invoke_operation(
-        "modular_form.level_one.named_q_expansion.compute",
-        {"form": "DELTA", "truncation_order": 4},
-        Catalog.open(),
-    )
-    assert result.output is not None
-    assert result.output["q_expansion"]["variable"] == "q"
-    assert result.output["q_expansion"]["coefficients"] == [
-        {"num": "0", "den": "1"},
-        {"num": "1", "den": "1"},
-        {"num": "-24", "den": "1"},
-        {"num": "252", "den": "1"},
-    ]
-
-
-def test_dispatch_output_series_feeds_existing_formal_series_operation_unchanged() -> (
-    None
-):
-    named = invoke_operation(
-        "modular_form.level_one.named_q_expansion.compute",
-        {"form": "E4", "truncation_order": 3},
-        Catalog.open(),
-    )
-    assert named.output is not None
-    squared = invoke_operation(
-        "formal_series.rational.power.compute",
-        {"series": named.output["q_expansion"], "exponent": 2},
-        Catalog.open(),
-    )
-    assert squared.output is not None
-    assert squared.output["result"]["coefficients"] == [
-        {"num": "1", "den": "1"},
-        {"num": "480", "den": "1"},
-        {"num": "61920", "den": "1"},
-    ]
